Comment, clean, refactor
