+++ internal/ask/form/form.go
@@ -1,10 +1,7 @@
 package form
 
 import (
-	"crypto/md5"
-	"encoding/hex"
 	"errors"
-	"fmt"
 	"time"
 
 	"github.com/ardanlabs/kit/log"
@@ -58,38 +55,9 @@ type Step struct {
 	Widgets []Widget `json:"widgets" bson:"widgets"`
 }
 
-// MCAnswerAggregation holds the count for selections of a single multiple
-// choice answer.
-type MCAnswerAggregation struct {
-	Title string `json:"answer" bson:"answer"`
-	Count int    `json:"count" bson:"count"`
-}
-
-// MCAggregation holds a multiple choice question and a map aggregated counts for
-// each answer. The Answers map is keyed off an md5 of the answer as not better keys exist
-type MCAggregation struct {
-	Question  string                         `json:"question" bson:"question"`
-	MCAnswers map[string]MCAnswerAggregation `json:"answers" bson:"answers"`
-}
-
-// TextAggregation holds the aggregated text based answers for a single question.
-type TextAggregation struct {
-	Question string   `json:"question" bson:"question"`
-	Answers  []string `json:"answers" bson:"ansewrs"`
-}
-
 // Stats describes the statistics being recorded by a specific Form.
 type Stats struct {
-	Responses        int                        `json:"responses" bson:"responses"`
-	MCAggregations   map[string]MCAggregation   `json:"mc_aggregations" bson:"mc_aggregations"`
-	TextAggregations map[string]TextAggregation `json:"text_aggregations" bson:"text_aggregations"`
-}
-
-// Group defines a key for a multiple choice question / answer combo.
-type Group struct {
-	QuestionID string `json:"question_id"`
-	Question   string `json:"question"`
-	Answer     string `json:"answer"`
+	Responses int `json:"responses" bson:"responses"`
 }
 
 //==============================================================================
@@ -319,231 +287,4 @@ func Delete(context interface{}, db *db.DB, id string) error {
 
 	log.Dev(context, "Delete", "Completed")
 	return nil
-}
-
-// MCAggregate calculates statistics on all multiple choice questions across submission
-// on a form. It only considers qustions in the current form as currently there is no
-// way to track how questions and answers change if the admin updates the form mid flight.
-// Aggregate returns a map[_question_]map[_choice_]int_count datastructure.
-func MCAggregate(context interface{}, db *db.DB, id string, subs []submission.Submission) (map[string]MCAggregation, error) {
-	log.Dev(context, "Aggregate", "Started : Submission[%s]", id)
-
-	if !bson.IsObjectIdHex(id) {
-		log.Error(context, "Aggregate", ErrInvalidID, "Completed")
-		return nil, ErrInvalidID
-	}
-
-	// Get the form in question.
-	form, err := Retrieve(context, db, id)
-	if err != nil {
-		return nil, err
-	}
-
-	// Create a container for the aggregations: [question][option]count.
-	aggs := make(map[string]MCAggregation)
-
-	// Find the MultipleChoice widgets and add them to the aggs map
-	for _, step := range form.Steps {
-		for _, widget := range step.Widgets {
-			if widget.Component == "MultipleChoice" {
-				aggs[widget.ID] = MCAggregation{
-					Question: widget.Title,
-				}
-			}
-		}
-	}
-
-	// In this section we are looking through all submissions for answers to multiple choice
-	// questions that are still active in the form and counting question/answer pairs.
-
-	// Look at all submisisons.
-	for _, sub := range subs {
-
-		// Then at every anwer.
-		for _, ans := range sub.Answers {
-
-			// Skip nil answers.
-			if ans.Answer == nil {
-				continue
-			}
-
-			// The following section points to the need to refactor form submissions / introduce
-			// stronger typing.
-
-			// Unpack the answer object.
-			a := ans.Answer.(bson.M)
-
-			options := a["options"]
-
-			// Options == nil points to a non MultipleChoice answer.
-			if options == nil {
-				continue
-			}
-
-			// This map of interfaces represent each checkbox the user clicked.
-			opts := options.([]interface{})
-			for _, opt := range opts {
-
-				// Unpack the option.
-				op := opt.(bson.M)
-
-				// Use the title of the option as the map key.
-				selection := op["title"].(string)
-
-				// Hash the ansewr text for a unique key, as no actual key exists.
-				hasher := md5.New()
-				hasher.Write([]byte(op["title"].(string)))
-				optKeyStr := hex.EncodeToString(hasher.Sum(nil))
-
-				// If this question is not in the map then we can skip as it is not a current answer.
-				if _, ok := aggs[ans.WidgetID]; !ok {
-					continue
-				}
-
-				// If this is the first answer for this question, make a map for it.
-				if aggs[ans.WidgetID].MCAnswers == nil {
-					tmp := aggs[ans.WidgetID]
-					tmp.MCAnswers = make(map[string]MCAnswerAggregation)
-					aggs[ans.WidgetID] = tmp
-				}
-
-				// If this is the first time we've seen this answer, init the agg struct for it.
-				if _, ok := aggs[ans.WidgetID].MCAnswers[optKeyStr]; !ok {
-					aggs[ans.WidgetID].MCAnswers[optKeyStr] = MCAnswerAggregation{
-						Title: selection,
-						Count: 0,
-					}
-				}
-
-				// Increment the counter for this question/answer pair.
-				tmp := aggs[ans.WidgetID].MCAnswers[optKeyStr]
-				tmp.Count++
-				aggs[ans.WidgetID].MCAnswers[optKeyStr] = tmp
-
-			}
-		}
-	}
-
-	log.Dev(context, "Aggregate", "Completed : Submission[%s]", id)
-	return aggs, nil
-}
-
-// TextAggregate returns all text answers flagged with includeInGroup.
-func TextAggregate(context interface{}, subs []submission.Submission) (map[string]TextAggregation, error) {
-
-	// Create a container for the aggregations: [question][option]count.
-	aggs := make(map[string]TextAggregation)
-
-	// Scan all the submissions and answers.
-	for _, sub := range subs {
-		for _, ans := range sub.Answers {
-
-			// Skip nil answers.
-			if ans.Answer == nil {
-				continue
-			}
-
-			// Only include answers of questions flagged with "includeInGroups".
-			props := ans.Props.(bson.M)
-			include, ok := props["includeInGroups"]
-			if include == nil || !ok {
-				continue
-			}
-
-			// Unpack the answer and append it to the aggregation slice for that WidgetId.
-			a := ans.Answer.(bson.M)
-			tmp := aggs[ans.WidgetID]
-			tmp.Question = ans.Question
-			tmp.Answers = append(tmp.Answers, a["text"].(string))
-			aggs[ans.WidgetID] = tmp
-
-		}
-	}
-
-	log.Dev(context, "Text Aggregate", "Completed : Submission")
-	return aggs, nil
-}
-
-// GroupSubmissions organizes submissions by Group. It looks for questions with the group by flag
-// and creates Group structs.
-func GroupSubmissions(context interface{}, db *db.DB, id string) (map[Group][]submission.Submission, error) {
-
-	if !bson.IsObjectIdHex(id) {
-		log.Error(context, "TextAggregate", ErrInvalidID, "Completed")
-		return nil, ErrInvalidID
-	}
-
-	// Get the submissions for the form.Collection
-	subs, err := submission.Search(context, db, id, 0, 0, submission.SearchOpts{})
-	if err != nil {
-		return nil, err
-	}
-
-	groups := make(map[Group][]submission.Submission)
-
-	// Scan all the submissions and answers.
-	for _, sub := range subs.Submissions {
-
-		// Add all submissions to the [all,all] group
-		group := Group{
-			Question: "all",
-			Answer:   "all",
-		}
-		tmp := groups[group]
-		tmp = append(tmp, sub)
-		groups[group] = tmp
-
-		// Look for answers that may contain sub groups.
-		for _, ans := range sub.Answers {
-
-			// Skip nil answers.
-			if ans.Answer == nil {
-				continue
-			}
-
-			// Only include answers of questions flagged with "includeInGroups".
-			props := ans.Props.(bson.M)
-			include, ok := props["groupSubmissions"]
-			if include == nil || !ok || include == false {
-				continue
-			}
-
-			fmt.Printf("\n\n%#v", props["groupSubmissions"])
-
-			// Unpack the answer object.
-			a := ans.Answer.(bson.M)
-
-			options := a["options"]
-
-			// Options == nil points to a non MultipleChoice answer.
-			if options == nil {
-				continue
-			}
-
-			// This map of interfaces represent each checkbox the user clicked.
-			opts := options.([]interface{})
-			for _, opt := range opts {
-
-				// Unpack the option.
-				op := opt.(bson.M)
-
-				// Use the title of the option as the map key.
-				selection := op["title"].(string)
-
-				// Add the submission to this subgroup
-				group := Group{
-					QuestionID: ans.WidgetID,
-					Question:   ans.Question,
-					Answer:     selection,
-				}
-				tmp := groups[group]
-				tmp = append(tmp, sub)
-				groups[group] = tmp
-
-			}
-
-		}
-	}
-
-	return groups, nil
 }
